requirements.txt
distribute==0.6.34

Django==1.6.6
South==1.0
-e git+https://github.com/hellsgate1001/django-audit-log@c4b88a14205820c741c257dbcd1899f0620bcde4#egg=django_audit_log-v1.0
django-braces==1.4.0
six==1.7.3
fake-factory==0.4.2
django-mptt==0.6.1

#Keep factory_boy at the bottom as there is dependency issues with distribute and setuptools
factory-boy==2.4.1
